refactor(InputParser): 패키지 변경
